server/pyproject.toml
[tool.poetry]
name = "sibdev deals test"
version = "0.1.0"
description = ""
authors = ["B4TRUH4 <matveybond2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.8"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
